backend done

+++ backend/controllers/product.controller.js
@@ -50,6 +50,10 @@ export const updateProduct = async (req, res) => {
 export const deleteProduct = async (req, res) => {
     const {id} = req.params;
 
+    if(!mongoose.Types.ObjectId.isValid(id)){
+        return res.status(404).json({sucess:false, message: 'Invalid product ID'});
+    }
+
     try {
         await Product.findByIdAndDelete(id);
         res.status(200).json({success: true, message: 'Product deleted successfully'});
